modules/locations_producer/requirements.txt
grpcio-tools==1.45.0
grpcio==1.45.0
kafka-python==2.0.2
